Stop socket accept error spew for lsp connections during init
Summary:
The `SocketAcceptor` was written to log a verbose error whenever it threw an exception. When `flow lsp` was built, it used this exceptional behavior to avoid setting up a connection during initialization. This led to some error spew.

This diff refactors the code a little bit to just log an `info` in that case.

Reviewed By: mroch

Differential Revision: D16841302

fbshipit-source-id: 005458222cdf3ae3a68606c60f349d9081ca5bb2