[jsx] Traverse props for namespaced name jsx
Summary:
Previously, we completely ignored jsx expressions with namespaced names.

This diff makes sure that we create a props object, which will find any errors that we would get from looking at the props themselves.

The examples in the added test case all previously did not error.

Next steps:

When using React, instantiate the jsx like a react component. To get the component config for `<a:b />` we need to look at `a.b`.

Reviewed By: gabelevi

Differential Revision: D16814363

fbshipit-source-id: 06d476196589b2d9785d62aa14386591b6cdf38e