partially fix autocomplete of object properties
Summary:
consider autocomplete inside an object literal, like this:

```
const x: T = {  };
              ^^ cursor between braces
```

we insert the autocomplete token `AUTO332` before parsing, so we actually parse `{ AUTO332 }`, which is a shorthand property, which is an identifier, so we return results for identifiers in scope, rather than for the valid properties of `T`.

this is almost certainly not what anyone wants. but it's a bit tricky right now to actually autocomplete the right properties, because of the way hooks fire while traversing the AST in statement.ml. we need to know the upper bounds that the literal's ObjT flows into, or match certain syntax like assignments, destructuring, and casts.

this diff makes us stop treating it as an identifier completion, and to return no results instead. but it also classifies it as an object key completion so we can measure how pervasive this is at the same time as we fix it properly.

Issue: https://github.com/facebook/flow/issues/7723

Reviewed By: panagosg7

Differential Revision: D16866760

fbshipit-source-id: b352fe3044a0fb3fe583dc11c0d362f792637224