[find-refs] hoist json logging out of FindRefs_js
Summary:
I think that specific logging formats should be handled by callers, leaving the find-refs service as agnostic as possible.

The 3 callers are `find-refs`, the LSP `textDocument/documentHighlight` and the LSP `textDocument/references`, and they all have slightly different logging requirements.

Additionally, I think we should call this from flow.js for Try Flow, but it wouldn't want to generate JSON, it would want to directly create JS objects.

Future work: now that it's clear that `documentHighlight` only logs `SUCCESS` vs `FAILURE`, it's also clear that it's redundant, since the presence of an `Error` key in the response does the same thing. but didn't want to break our existing metrics in this diff.

Reviewed By: gabelevi

Differential Revision: D16987827

fbshipit-source-id: ed4a038a1eee7ca979f825a18fa2602a6c94cebf