insert_type_utils validator returns a list of errors instead of throwing
Summary:
The `type_validator_visitor` of insert_type_utils.ml used to throw an exception on the first
validation error it would encounter. For clients that need to insert some version of this type,
this meant that the output had to be a very general default type (usually `any`). It is often
desirable, however, to return a more detailed version of the type that only has `any` in those
places that cause validation failures. This diff populates a list of errors gathered as we validate
types that are candidates for insert_type.

Reviewed By: mroch

Differential Revision: D16774596

fbshipit-source-id: d5e87bde14e645aab11c5fe38169d47b70fa9dd3