[PR] Add [rollouts] to .flowconfig
Summary:
## Motivation

We build perf improvements into Flow all the time, but it's a little hard to verify the effect in practice when you're comparing Flow usage this week to Flow usage last week. There are a lot of confounding factors that can pop up.

Instead, we can use rollouts to turn a feature on for X% of the users. This will lead to better numbers.

## Syntax

There were a few tricky things that I wanted to capture in the syntax

1. Options currently have the form `IDENTIFIER=VALUE`, where `IDENTIFIER` is `[a-zA-Z0-9._]+` and `VALUE` is `.*`. This makes it tricky to add syntax to the RHS of an option, since the syntax could interfere with an option. I'd much rather add syntax to the LHS since the LHS is already more restricted
2. It would be nice if you tie multiple options together. Like have either (`foo=A` and `bar=A`) or have (`foo=B` and `bar=B`)
3. Some options support multiple entries (e.g. `module.system.haste.paths.blacklist`). It would be nice to support this too
4. It would be nice if we could also rollout changes to non-options settings too.

So the syntax I came up with had two parts

1. A `[rollouts]` section where you define a rollout, its groups, and each group's probability
2. A way of including lines of the .flowconfig based on the rollout group

The syntax for #1 is pretty familiar

```
[rollouts]
project_doomsday=50% on, 50% off
when_to_drop_the_beat=50% now, 10% tomorrow, 40% next_week
```

And then there's a prefix syntax where a line is only used if the specified rollout is in the specified group.

```
[options]
(project_doomsday=on) go_boom=true ; This line is included when project_doomsday chooses the "on" group
(project_doomsday=false) go_boom=true ; This line is included when project doomsday chooses the "off" group
```

## Bucketing

We want the rollout bucketing to be somewhat consistent. Therefore, the pseudo-random number generator is keyed off 3 things:

1. The rollout's name
2. The hostname
3. The user's ID

That should make the rollout bucketing consistent enough between runs not to cause much trouble.

 ---

Pull Request resolved: https://github.com/facebook/flow/pull/8018

Test Plan:
I added a unit test which tests the deterministic parts of this feature, like parse errors and setting a rollout group to 0% or 100%.

In `FlowConfig.calculate_pct` I changed the code to

```
  let ret = Xx.modulo hash 100 in
  Printf.eprintf "Rollout %S: my pct is %d\n%!" rollout_name ret;
  ret
```

then when I ran the test I saw the following related lines

```
...
+Rollout "foo": my pct is 10
...
+Rollout "foo": my pct is 10
...
+Rollout "foo": my pct is 10
+Rollout "foo": my pct is 10
...
+Rollout "foo": my pct is 10
...
+Rollout "rollout": my pct is 34
...
+Rollout "verify_sig": my pct is 88
...
```

Note that the same rollout names always have the same values.

To test the logging I modified `packages/flow-dev-tools` and added

```
[rollouts]
gabe_test_1=50% on, 50% off
gabe_test_2=10% blue, 60% red, 30% green
```

Then I

  ptail -f -i perfpipe_flow_server_events | grep "$HOSTNAME"

and ran

  flowd check packages/flow-dev-tools

In the `INIT_END` event I saw

```
"tags":{"rollouts":["gabe_test_2=red","gabe_test_1=off"]}
```

Reviewed By: mroch

Differential Revision: D16786071

fbshipit-source-id: 09abf091dea8f11ef87778c5f87a832ad508e13a