[find-refs] split into helper functions
Summary:
refactors `find_refs` to be slightly more "linear". previously, `property_find_refs` was always called, but the loc it was called with depended on whether local find refs worked.

the overall goal here is to split `find_refs` into separate local and global versions later in the stack, and to minimize the amount of repeated code.

Reviewed By: nmote

Differential Revision: D17057696

fbshipit-source-id: 74e7b746872f4bbaa40fa29390e8bed3ef3f1866