ShapeT should be compatible with ExactT
Summary:
Compatibility rules for `ShapeT` as lower bounds are currently very loose: `ShapeT(o)` can be used however `o` can be used.

However, because the compatibility rules for `ExactT` currently precede those for `ShapeT` in our implementation, and because `ExactT` errors using a wildcard, `ShapeT` is currently incompatible with `ExactT`.

This does not appear to be by design. For example, if a `ShapeT` flows to a `ExactT` through a `ReposLowerT`, there is currently no error.

Therefore this diff makes `ShapeT` explicitly compatible with `ExactT`. In the future, we might harden `ShapeT` but that's a big project.

Reviewed By: panagosg7

Differential Revision: D17089642

fbshipit-source-id: 381cfcaba8736e21cb71f9e3ccda64a0f97a772d