add explicit flag to replace_reason
Summary:
We're finally ready to split replace_reason into two functions, one that preserves def_loc (derived reason), and the other that doesn't (new reason).

One complication is that both versions keep annot_loc but I'm going to fix that in a later diff. I want def_loc and annot_loc to be paired, and annot_loc to be set explicitly where needed.

Reviewed By: panagosg7

Differential Revision: D17054314

fbshipit-source-id: 444d346e6f3a6225aefa034edbc6aaa68de4c6f1