fix unexpected internal type exception
Summary:
Internal types are not supposed to become use types, and violating this leads to an assertion failure. Unfortunately internal types can hit `AnnotT`, which wraps them in a `ReposUseT`, which can then be exposed as use types in a rule involving `MergedT`.

But `MergedT` (1) cannot arise in types-first and (2) without types-first, can hit `ReposUseT` leading to a no-op at best (flowing a type to a repositioned version of itself) and a crash because of the above problem at worst. So this diff deletes the `MergedT ~> ReposUseT` rule (until we delete `MergedT` completely).

Reviewed By: panagosg7

Differential Revision: D17084689

fbshipit-source-id: f67f6f6760f1e259e2b8f736101bfc09037fd7ec