[enums] Remove usage of lookahead for member init
Summary: In our original approach to parsing member inits, if we had an invalid initializer we wanted to parse the entire thing to get the location for the whole invalid init. To do this we required the lookahead. However, we changed our apporach to invalid inits to not parse the entire thing, so the lookahead is not necessary. In general, we should avoid lookaheads when possible.

Reviewed By: gabelevi

Differential Revision: D17136550

fbshipit-source-id: bb9eb58106842e8f2070b886eae792de8e2fadd2