Avoid repetition of extends type in class sig
Summary:
In Class_sig.supertype we create a type representing the extended class. We end
up doing the same work again in Class_sig.statictype, since we want the statics
of the super class.

This change avoids the redundant specialization. Note that this is closer to
what we already do in the Class_sig.toplevels logic.

Reviewed By: avikchaudhuri

Differential Revision: D16878280

fbshipit-source-id: 0be09dccd2c19118444d8ac7d20c96fcacfb2761