[autocomplete] Improve behavior of extract_type by removing implicit preconditions
Summary:
The `extract_type` function in members.ml, which is used in autocomplete, has several unenforced preconditions on its input: it assumes that, if necessary, `resolve_type` and `resolve_builtin_class` have already been called on its argument if necessary, and if that is not the case, it returns `FailureUnhandledType`, which leads to an autocomplete failure with an unexpected type. This precondition did not always hold: some recursive calls of `extract_type` neglected to call these functions on their arguments. For example, autocomplete would fail on values of type `?Array<number>`, because `Array<number>` isn't resolved before its type is extracted.

This diff removes those preconditions by having `extract_type` call `resolve_type` when necessary, and folding the behavior of `resolve_builtin_class` into `extract_type` entirely (because that function is not used anywhere else). This fixes a source of unexpected type errors* during autocomplete.

This diff also adds a new kind of failure that can be generated by `extract`. Currently, there's no way to tell whether a `FailureUnhandledType` originates in `extract_type` (if the type doesn't have a corresponding object-like type containing its members) or if in `extract_members` (if the provided object-like type doesn't actually contain members). This diff adds `FailureUnhandledMembers` for the latter case, which should help debug future autocomplete failures.

Reviewed By: mroch

Differential Revision: D17141837

fbshipit-source-id: 36e9c4fa4ebb6c3d012ff2b8a97cefb6d4b9994f