Tests for recheck behavior when module location changes
Summary:
This tests what happens when a file is added/removed in a way which affects which file a module resolves to in another file.

When we start skipping direct dependents of changed files (as much as recheck opts allow), we want to make sure that this behavior is correctly maintained.

Reviewed By: avikchaudhuri

Differential Revision: D16725172

fbshipit-source-id: e5df70ad14205f56be17e0391c12d71d66a3d5ed