Remove read from global state in `focused_files_and_dependents_to_infer`
Summary:
I'm not a huge fan of this change, but I didn't see a better alternative which would allow us to remove the read from global state from this function.

I investigated what it would take to remove this check entirely, but it's not clear what it would take. I initially tried skipping it, and all the tests pass. Then I tried having it raise an exception instead of filtering a file, and only a handful of LSP tests failed. This led me to the conclusion that it's currently at least doing *something*, even if it's not very important.

Reviewed By: gabelevi

Differential Revision: D17192001

fbshipit-source-id: 2e730337d160b9eb040ad4620fe9724760995195