Remove `env` from `determine_what_to_recheck` function
Summary: This change improves this function's testability. We only need a couple things out of the `env`, so let's pass them in directly. We also have to read global state to get the files that the IDE has open, so extracting this read to the caller makes the function closer to pure.

Reviewed By: gabelevi

Differential Revision: D16990466

fbshipit-source-id: 80835846ec34b8043a537653430e629e858103c8