[dune] fix fsnotify dune file on Windows
Summary:
on Windows (64-bit), the "system" is one of `mingw64`, `win64` or `cygwin`, depending on whether it was built with mingw, MSVC, or cygwin respectively. the Windows version of `fsnotify` is `fsnotify_Windows`, so the previous fallback didn't work.

also, since there are no fallbacks for other platforms, I made it not generate the `fsnotify` virtual library at all in that case. that way, you end up at `fsnotify/dune` to figure out what's wrong, instead of looking for something bogus like `fsnotify_beos`.

Reviewed By: gabelevi

Differential Revision: D17261740

fbshipit-source-id: 1844fef0bc427af08222a056379eccf953525f1d