[hack] fix lz4 include paths on Windows
Summary:
cmake sets some environment variables to tell dune where to find external libraries like lz4. sometimes there's nothing to set so we set an empty string.

on linux/mac, `Sys.getenv` delegates to C's [`getenv`](http://man7.org/linux/man-pages/man3/getenv.3.html) which returns `""` for set-but-empty, or null when the env is unset (which raises `Not_found`).

on Windows, `Sys.getenv` calls [`GetEnvironmentVariable`](https://docs.microsoft.com/en-us/windows/win32/api/winbase/nf-winbase-getenvironmentvariable), which returns the number of bytes read, or 0 when the env is not found. an empty string is also 0 bytes, but ocaml doesn't check `GetLastError` to distinguish the two cases. I think this is an ocaml bug.

in this case, I don't think we really care if the env is set or not, so just made it handle `""` and `Not_found` the same. since we don't need the `Option` anymore, it simplifies a lot of stuff.

Reviewed By: fredemmott

Differential Revision: D17266309

fbshipit-source-id: 16c483f8586f781a02cd7a3bb0681cfa307b11c1