[hack] fix build-timestamp-opt on Windows
Summary:
we run a bash command to compute the build timestamp. the `system` action runs `sh` on linux and `cmd` on Windows, but since this is a bash script, we need to always use `bash`. luckily, we build for Windows under cygwin so this works on Windows too.

this appears to be the Windows equivalent of `$()`, if we wanted to write a batch script that `system` could call: https://stackoverflow.com/a/2768658. if we have to stop using `bash` because of, e.g. Linux distros that ship without bash, then we can figure out the batch script.

Reviewed By: gabelevi

Differential Revision: D17269476

fbshipit-source-id: 8be29db4b80dc44e160d4d52dba0e0c0fc1d9656