Refactor lsp diagnostic handling into LspErrors module
Summary:
The way that the Flow lsp deals with diagnostics (aka errors) is kind of complicated. Errors come from multiple sources, are stored in various places, and are combined in different ways depending on the circumstances. This diff tries to pull all of that into the `LspErrors` module. The goals of this diff are

1. Clean up and consolidate the code
2. Make it unit testable
3. Make it easy in the future to add another source of errors (running `check-contents` on change)

 ---

The current status quo is this:

1. *live parse errors*. If live errors are turned on, then the Flow lsp process will parse a file on change in order to figure out the parse errors. This works even if the Flow server is not connected. If we also have Flow server errors for this file, then we'll combine our live parse error with the server's non-parse errors. These we previously stored in the `open_file_info` structure.
2. *streamed errors*. Errors that come from the server during a recheck are considered to be additive. These were stored in the `connected_env`
3. *finalized errors*. Errors that come from the server outside of a recheck. They are considered complete and will replace existing server errors. Files not mentioned are considered to have 0 server errors. These were also stored in the `connected_env`

In general, we try not to send messages to the client when it's obvious that the errors for a file have not changes (e.g. 0 errors -> 0 errors).

Reviewed By: mroch

Differential Revision: D16907166

fbshipit-source-id: 36393c82986d611dd52e7dbdc5cce66d83e7f2bd