add Pointer Lock libdefs
Summary:
adds the missing parts of the [Pointer Lock](https://w3c.github.io/pointerlock/) spec:

- `document.pointerLockTarget`
- `document.exitPointerLock()`

didn't add 'pointerlockchange' and 'pointerlockerror' to `addEventListener` because they don't seem to create a special kind of event, so can use the fallback event handler. but we could reconsider that in a followup to improve autocomplete.

Reviewed By: gabelevi

Differential Revision: D17285731

fbshipit-source-id: e29617d266517357447188693ab878ce825aaf00