[hack] Avoid calculating code digest from top level
Summary:
In D13206371, I fixed an inefficiency where forked job processes needed to
calculate the digest of the code segment. That change added a `Marshal.to_bytes`
call to the top level. This causes any binary which includes the Worker module
to invoke the `to_bytes` logic at startup.

Since every command uses the same binary, even commands like `flow version` pay
this cost. I outlined the consequences of that more in D17299048.

This diff moves the `to_bytes` call into the worker pre-fork code as well, so we
only pay the cost of computing the digest from workers instead of always.

Reviewed By: gabelevi

Differential Revision: D17299468

fbshipit-source-id: ce0793eb987eb904bef737402752b74df367f786