Test to show how batch-coverage is broken
Summary:
`flow batch-coverage` throws away all coverage info whenever we recheck, replacing it with the coverage info for the files that we recheck. So if you have 10 files and recheck 1 of them, suddenly you only have coverage info for 1 file.

This doesn't fix it, it just adds a test to demonstrate the problem.

Reviewed By: mroch

Differential Revision: D17399743

fbshipit-source-id: 751e272d6260e8d1b9517942499aa41a3640c450