Test to show how batch-coverage is broken
Summary:
`flow batch-coverage` throws away all coverage info whenever we recheck, replacing it with the coverage info for the files that we recheck. So if you have 10 files and recheck 1 of them, suddenly you only have coverage info for 1 file.

This doesn't fix it, it just adds a test to demonstrate the problem.

Reviewed By: mroch

Differential Revision: D17399743

fbshipit-source-id: b17f2f9f9cfa3b2527e2b6158adc638de90c4f08