Use RelativeLoc in ALoc table
Summary:
This leads to a minor memory win when abstract locations is turned on. This is a constant size (not percent) win no matter whether lazy mode is on or off, because all files are parsed, and therefore abstract locations tables stored, even if not all files are checked due to lazy mode.

I ran experiments in the following four configurations, and this one was the best:

* Store ordinary locations in the tables (behavior before this change)
* Store relative locations with sources (surprisingly, this was actually a regression)
* Store relative locations without sources (this diff, best configuration)
* Store a variant of ordinary location that does not include the source field (pretty good, but slightly worse than this diff)

Reviewed By: mroch

Differential Revision: D16359026

fbshipit-source-id: fea4da3d3ff29ee9c3bf329fd31b7697365300be