Log PushedLiveParseErrors user experience to flow_interactions
Summary: It would be good to have visibility into how quickly and reliably we display live syntax errors to the user. This adds logging to `flow_interactions` whenever we send live syntax errors to the user. This lets us see things like "how fast after a didChange did we send the live syntax errors to theuser".

Reviewed By: mroch

Differential Revision: D17374349

fbshipit-source-id: 09c94cdc5aacdf5478e70b57727b979c66539ec2