Don't set live parse errors in do_documentSymbol
Summary:
There's two reasons that the Flow LSP process will parse a file

1. It was opened or changed. We parse it in order to see if there are syntax errors, so that we can quickly tell the client, even if the Flow server isn't connected
2. We get a documentSymbol request. We parse the AST in order to return a list of symbols in the document. This also works  without a Flow server

In order to avoid reparsing the same file, we cache the file.

Previously, documentSymbol requests could in theory send live parse errors to the client. In practice, we probably receive the didOpen/didChange before the documentSymbol request, so this likely wasn't happening, but it's possible from the code.

After this diff, we'll ONLY send live parse errors immediately after a didOpen/didChange/didSave.

On the off chance that a documentSymbol parses a file (and doesn't just read the cached AST), we now also cache the parse errors.

Reviewed By: mroch

Differential Revision: D17353329

fbshipit-source-id: dac9e842e422c870bda1112487d0ae39d557c129