improve error message for value used as type
Summary:
Previously, a value used as a type would produce an error that only contained the description of the value.

Instead, this diff makes the error contain a description of the use and a reference to the value, which adds some useful context to the error. Added some more wording, too.

Reviewed By: samwgoldman

Differential Revision: D17269890

fbshipit-source-id: 526b3901a740e1a31c2ddec6d53983280dce15ac