hack: Pass worker_id to multiWorker jobs
Summary:
merge now passes the worker_id into the merge step. I'd like similar
functionality for the job function as well. I tried inferring this information
other ways, but it ends up being really messy.

Similar to with the merge step, we provide a version of the calls without the
worker_id passed in, for the existing call sites where the id isn't needed.

Reviewed By: fahndrich

Differential Revision: D17407537

fbshipit-source-id: 251f958c3a62201efb28a092e899a016730c078b