Split the server -> lsp messages into responses and notifications
Summary:
Every message sent from the Flow server to the Flow LSP is either

1. A response to a request
2. A notification, which isn't a response to anything in particular

Previously they were mixed together in a single type. This diff splits them up. This allows me to force `commandHandler.ml`'s code to always return a response and never a notification. This also lets me enforce that a response is never sent when a notification is appropriate.

Reviewed By: mroch

Differential Revision: D17463326

fbshipit-source-id: 3ff8cd5fb49ac21b1a3aaee346d803765f337274