Move LSP's Subscribe request to LspResponse
Summary:
When `commandHandler.ml` handles a request from a persistent client, it returns either `IdeResponse` or `LspResponse`. Now that `flow ide` is dead, I want to kill `IdeResponse`. The only thing that still uses it is the `Subscribe` message, mainly because it lacks `metadata` so can't send an `LspResponse`.

The easy fix is to add `metadata` to the request so the response can use `LspResponse`

Reviewed By: mroch

Differential Revision: D17453297

fbshipit-source-id: 1edee8a199de18b8ed8cf1e2b23d31bf9c105bd1