Implement checking and refinements for delete
Summary:
Implements logic for member and variable deletion.
The basic goal of this diff is to prevent this kind of unsoundness:
```
declare var x: {foo: number}
delete x.foo;
(x.foo: number) // Unsound!
```
This diff accomplishes that by raising an error on `delete x.foo`; on the other hand, if `x` was defined as `{foo?: number}`, the deletion would be ok.

For the most part, deleting a member or element can be modeled just as flowing `VoidT` to that member or element, and that is all that is needed for cases like the above example. Unfortunately, it's more complicated for objects with indexers:
```
declare var x: {[string]: number}
delete x.foo // arguably, should be fine
```
Objects with indexers are already unsound with respect to undefined, so it would be inconsistent to error when we delete something from them. To allow for this behavior (and analogous behavior for arrays) to be accepted, we unfortunately need to thread an extra bit of information through `SetPropT` and `SetElemT`, which specifies whether the type represents a regular member/element assignment or a deletion. (We could create new types for `DeletePropT` etc, and it could be done if that's preferred, but the behavior of deletion and assignment is *almost* the same, so I went with using the same types but adding an extra parameter.) We also thread some information back from property lookups, to represent whether the lookup succeeded via an indexer or otherwise, and when a deletion hits an indexer prop, we don't flow `VoidT` to the prop like we otherwise would.

Since we're mostly just reusing the member assignment machinery, refinements work out of the box, and we error when deleting from read-only props, `$ReadOnlyArray`s, etc.

Reviewed By: panagosg7

Differential Revision: D16867974

fbshipit-source-id: 9d504f5b2bc8e42e6cf871fedfc1fdf943623aa7