Better handling of Object.defineProperty and friends
Summary:
Currently Flow special-cases `Object.defineProperty`, `Object.defineProperties`, and `Object.create` in ways that don't correspond with their specifications or with their libdef definitions in `core.js`. Each of these methods take `PropertyDescriptors`, which are defined in spec as
```
type PropertyDescriptor<T> = {
    enumerable?: boolean,
    configurable?: boolean,
    writable?: boolean,
    value?: T,
    get?: () => T,
    set?: (value: T) => void,
    ...
};
```
But the special-cased implementation of these methods requires that they have a `value` property, whose type is used to initialize the property on the Flow object type that is being created or modified. The rest of the properties are ignored--which means that the `get` property can return something of an entirely different type.

This diff alters that behavior to correspond to the spec by actually looking up the definition of `PropertyDescriptor` and flowing the arguments of these `Object` methods into that definition. The methods are still special cased, but instead of requiring a `value` property contrary to spec, the type for the object property is determined by creating a tvar `X` and then flowing the property descriptor argument into `PropertyDescriptor<X>`. `X` then becomes the type on the ObjT for the property. This corresponds to spec and improves soundness, because now if the type of `value` and the type of `get` disagree, an error will be raised.

There is also some additional work to correspond to spec by allowing `Object.defineProperty` to take a type argument, which will be used instead of the tvar above. Finally, `Object.freeze` is also changed to allow a type argument, as allowed in the spec.

Reviewed By: panagosg7

Differential Revision: D17081593

fbshipit-source-id: d7b04582a0ebc74204381caf7d23c30c569169a2