Remove files which lose the @flow pragma from the heap
Summary:
We don't add context heap entries for files without `flow`. However, if a files loses the `flow` pragma, we also don't remove its heap entries.

This behavior can lead to some small amount of unnecessary memory usage, but more importantly the stale data leads to incorrect results when the optimizations in D17552054 are applied.

D17535894 adds a test which fails after D17552054, when this diff is not applied. With the fix in this diff, the test passes even when we start allowing direct dependents to be skipped during check.

Reviewed By: gabelevi

Differential Revision: D17537927

fbshipit-source-id: f736a082afa18748e4ee8392170fe39827c9c053