Skip direct dependents in check
Summary:
This allows the check phase (applicable to types-first only) to skip checking direct dependents of files where the signature has not changed. Previously, we never applied recheck opts to direct dependents.

In order for this to be correct, I had to disable the optimization in several non-common cases. If the dependency graph changes or a file has a parse error, we don't apply recheck opts to direct dependents of changed files. This is still more conservative than we need to be, but I think this change as-is will capture the vast majority of the optimization opportunity.

Reviewed By: gabelevi

Differential Revision: D17552054

fbshipit-source-id: 1b7c8151b0c21cebeb529e41dec894f84a13e7d5