Incremental test for removing and later adding the @flow pragma
Summary:
This ensures correct recheck behavior when `flow` is removed and then later added back to a file. This catches an issue that I found when attempting to allow recheck opts to apply to direct dependents of the changed file(s).

This test fails when run against my current work in progress on skipping direct dependents. I'll fix the underlying issue in a future diff, which I will land before my work on skipping direct dependents.

Reviewed By: avikchaudhuri

Differential Revision: D17535894

fbshipit-source-id: 37785a94e1ccf32bcd6f68819b6ceaceb060d323