Enhance test for skipping files during rechecks
Summary: This is basically just here to validate my assumptions about what the "unparsed set" in `types_js.ml` is. It appears to be the set of files involved in the current recheck which are unparsed. I wanted to make sure it wasn't the set of all unparsed files in the repository. If it were, this test would stymie the recheck optimizations on direct dependents which I will add later in the stack.

Reviewed By: avikchaudhuri

Differential Revision: D17537258

fbshipit-source-id: 878e270e318a6debeb505a323239dc21be2d14d6