fix(declarations): suppress errors after merge, not during
Fixes #6631, #6717, #7803.

There are two major flaws in the original approach:

1. Silencing errors at the merge site does not work for builtins,
as those errors are computed later, and
2. Putting that suppression in an if/else where the else block
parsed comment error suppressions can cause [declarations] to
actually create *more* errors, which was misleading a lot of people
and masked the source of bugs.

We now suppress entire files in error collation, at the very end
of the error reporting process.