Unbreak abstract-locations-minimal-merge test under saved state
Summary: For some reason, the file removed in this diff gets skipped when we hydrate from saved state, but not normally. Getting rid of it does not significantly alter the character of this test, so for now I'm just doing that. We should probably spend some time on understanding why this file gets skipped when the server is started from saved state, and not normally.

Reviewed By: dsainati1

Differential Revision: D17641813

fbshipit-source-id: 733b59f0165a645bfb47db692c97881ef85ad33f