[spreads] Move Cache and instantiation utils into their own files
Summary:
These modules are taking up space in flow_js.ml and do not need to be.

I want to move `Cache` out of `flow_js.ml` so I can use it for my spread work (and avoid creating a dep cycle), but that module depends on `ImplicitTypeArgument` and `TypeAppExpansion`, so I moved them out as well.

Reviewed By: gabelevi

Differential Revision: D16674761

fbshipit-source-id: adffd4c3dc08f9d2c2d498352868d94ee0170443