[spreads] Add cache to prevent exponential explosion (Likely to cause new errors)
Summary:
This description is long because I want to page everything back in that we were discussing last half.

There's something I'm unsure about, which I marked with `[1]`.

# Preventing Exponentially Large Spreads

There are three ways to create an exponentially large type via spread. This diff addresses one. I will follow up on this diff with fixes to the remaining causes.

If a spread has `n` many operands, each of which has  `m` lower bounds, then the result of the spread operation can be up to `m^n` in size. In order to protect against that, we are limiting the amount of operands that may have multiple lower bounds in a spread to just 1. Enforcing this invariant is important for moving our value spread implementation over to the type spread mechanism.

## This diff in some detail

In order to enforce this invariant, we add a cache that keeps track of lower bounds added to a spread. Whenever a spread operand is resolved, we add the slice to a set for the slices we've seen at that index.

If we notice that we have multiple indices with more than one lower bound, we emit an error and stop resolving the spread.

## But should we use this for type spread?
Note that I create the `spread_id` that maps back to this information in the cache in `eval_destructor`. I can't create the `spread_id` for the syntactic spread because that may be part of a polymorphic signature. Instead, I need to wait until that polymorphic signature is being used to compute a type.

This makes a few funky edge cases, which are demonstrated in the tests. For example, UnionTs push down the destructor into their members for speculation purposes. That means that if the last spread operand (which triggers the computation) is a union, then we will create a new spread_id for each element of the union. This weakens our linear guarantee in type spread to only be a quadratic guarantee instead.

[1] I am not sure how to fix that. I considered assigning a spread_id to the destructor during substitution, but it's not safe to look behind OpenTs for EvalTs at that point. In general, anything in our type system that would cause multiple EvalTs to be triggered for the same syntactic occurrence of this destructor will suffer from this problem with the implementation here.

The good news is that this won't apply to value spread, which won't be using the EvalT mechanism. Since value spreads can't be polymorphic types (by definition), we can safely assign the spread_id at the syntactic spread.

That makes me question if these errors are even worth emitting for type spread, since
we can't provide the same guarantees as we do for value spread, and since these errors are emitted so infrequently.

Reviewed By: avikchaudhuri

Differential Revision: D16674760

fbshipit-source-id: 5ebacad709ed5bd98a2174ad863ca04d1e5c5cfb