Restore abstract-locations-minimal-merge test
Summary:
This reverts the changes made in D17641813. This test was failing under saved state for unknown reasons, so I made a minimal modification to unbreak the tests.

I've since investigated, and discovered that saved state does not save the sig_file_sig, which is used (under types-first) to construct the sig dependency graph. This means that when a Flow server is initialized from saved state, merge is driven using the implementation dependency graph, changing what files are rechecked and when. To my knowledge this cannot lead to incorrect results, but it does come with a performance cost.

Gabe would prefer to leave this behavior as-is right now, because minimizing saved state size is a priority. He said he would look into this issue before launching saved state.

Because my recheck tests are meant to test specifically which files are rechecked in response to which events, this issue means that there will often be discrepancies between their output when they are run normally and when they are run from a server initialized from saved-state. Because of that, I'm skipping saved state for this test and future tests of this nature.

Reviewed By: gabelevi

Differential Revision: D17690800

fbshipit-source-id: f561a677ca832167c7e3543fa7adb6fcb6a5134f